Created and added CustomHttpException to Tenant, Search and Event Service

+++ backend/scm/src/main/java/com/scm/scm/predefinedSearch/services/PredefinedSearchServices.java
@@ -2,6 +2,8 @@ package com.scm.scm.predefinedSearch.services;
 
 import com.scm.scm.predefinedSearch.dao.PredefinedSearchRepository;
 import com.scm.scm.predefinedSearch.vao.PredefinedSearch;
+import com.scm.scm.support.exceptions.CustomHttpException;
+import com.scm.scm.support.exceptions.ExceptionCause;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,10 +18,10 @@ public class PredefinedSearchServices {
 
     public PredefinedSearch addPredefinedSearch(PredefinedSearch predefinedSearch) {
         if (predefinedSearchRepository.existsById(predefinedSearch.getSearchBy())){
-            throw new RuntimeException("Predefined search already exists");
+            throw new CustomHttpException("PredefinedSearch with this id already exists", 400, ExceptionCause.USER_ERROR);
         } else {
             if (Objects.equals(predefinedSearch.getTitle(), "")) {
-                throw new RuntimeException("Predefined search title is empty");
+                throw new CustomHttpException("PredefinedSearch title is empty", 400, ExceptionCause.USER_ERROR);
             } else {
                 predefinedSearch.setId(predefinedSearch.generateId(predefinedSearch.getTitle()));
                 predefinedSearchRepository.save(predefinedSearch);
@@ -33,7 +35,7 @@ public class PredefinedSearchServices {
     }
 
     public PredefinedSearch getPredefinedSearchById(String id) {
-        return predefinedSearchRepository.findById(id).orElseThrow(() -> new RuntimeException("PredefinedSearch not found"));
+        return predefinedSearchRepository.findById(id).orElseThrow(() -> new CustomHttpException("PredefinedSearch not found", 404, ExceptionCause.USER_ERROR));
     }
 
     public PredefinedSearch updatePredefinedSearch(PredefinedSearch predefinedSearch) {
@@ -47,22 +49,34 @@ public class PredefinedSearchServices {
             oldPredefinedSearch.setFilter(predefinedSearch.getFilter());
             predefinedSearchRepository.save(oldPredefinedSearch);
         } else {
-            throw new RuntimeException("PredefinedSearch not found");
+            throw new CustomHttpException("PredefinedSearch not found", 404, ExceptionCause.USER_ERROR);
         }
         return oldPredefinedSearch;
     }
 
     public String deletePredefinedSearch(String id) {
-        predefinedSearchRepository.deleteById(id);
-        return "PredefinedSearch successfully deleted";
+        if (predefinedSearchRepository.existsById(id) && !id.isEmpty()) {
+            predefinedSearchRepository.deleteById(id);
+            return "PredefinedSearch successfully deleted";
+        } else {
+            throw new CustomHttpException("PredefinedSearch not found", 404, ExceptionCause.USER_ERROR);
+        }
     }
 
     public List<PredefinedSearch> getPredefinedSearchByUser(String user) {
-        return predefinedSearchRepository.findByUser(user);
+        if (!user.isEmpty()){
+            return predefinedSearchRepository.findByUser(user);
+        } else {
+            throw new CustomHttpException("PredefinedSearch user is empty", 400, ExceptionCause.USER_ERROR);
+        }
     }
 
     public List<PredefinedSearch> getPredefinedSearchByTenant(String tenant) {
-        return predefinedSearchRepository.findByOnTenant(tenant);
+        if (!tenant.isEmpty()){
+            return predefinedSearchRepository.findByOnTenant(tenant);
+        } else {
+            throw new CustomHttpException("PredefinedSearch tenant is empty", 400, ExceptionCause.USER_ERROR);
+        }
     }
 
 }
